+++ rasa/nlu/data_router.py
@@ -15,7 +15,7 @@ from rasa.nlu.components import ComponentBuilder
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.emulators import NoEmulator
 from rasa.nlu.test import run_evaluation
-from rasa.nlu.model import InvalidProjectError
+from rasa.nlu.model import InvalidProjectError, UnsupportedModelError
 from rasa.nlu.project import (
     Project, STATUS_FAILED, STATUS_READY, STATUS_TRAINING, load_from_server,
     FALLBACK_MODEL_NAME)
@@ -36,16 +36,18 @@ logger = logging.getLogger(__name__)
 DEFERRED_RUN_IN_REACTOR_THREAD = True
 
 
-class MaxTrainingError(Exception):
-    """Raised when a training is requested and the server has
-        reached the max count of training processes.
+class MaxPoolProcessError(Exception):
+    """Raised when training or evaluation is requested and the server has
+        reached the max count of pool processes.
 
     Attributes:
         message -- explanation of why the request is invalid
     """
 
     def __init__(self):
-        self.message = 'The server can\'t train more models right now!'
+        self.message = ("The server has reached its limit on process pool "
+                        "workers, it can\'t train or evaluate new models "
+                        "right now")
 
     def __str__(self):
         return self.message
@@ -82,15 +84,15 @@ def deferred_from_future(future):
 class DataRouter(object):
     def __init__(self,
                  project_dir=None,
-                 max_training_processes=1,
+                 max_pool_processes=1,
                  response_log=None,
                  emulation_mode=None,
                  remote_storage=None,
                  component_builder=None,
                  model_server=None,
                  wait_time_between_pulls=None):
-        self._training_processes = max(max_training_processes, 1)
-        self._current_training_processes = 0
+        self._pool_processes = max(max_pool_processes, 1)
+        self._current_pool_processes = 0
         self.responses = self._create_query_logger(response_log)
         self.project_dir = config.make_path_absolute(project_dir)
         self.emulator = self._create_emulator(emulation_mode)
@@ -111,7 +113,7 @@ class DataRouter(object):
         # -258934405
         multiprocessing.set_start_method('spawn', force=True)
 
-        self.pool = ProcessPool(self._training_processes)
+        self.pool = ProcessPool(self._pool_processes)
 
     def __del__(self):
         """Terminates workers pool processes"""
@@ -289,8 +291,8 @@ class DataRouter(object):
         # be other trainings run in different processes we don't know about.
 
         return {
-            "max_training_processes": self._training_processes,
-            "current_training_processes": self._current_training_processes,
+            "max_pool_processes": self._pool_processes,
+            "current_pool_processes": self._current_pool_processes,
             "available_projects": {
                 name: project.as_dict()
                 for name, project in self.project_store.items()
@@ -308,8 +310,8 @@ class DataRouter(object):
         if not project:
             raise InvalidProjectError("Missing project name to train")
 
-        if self._training_processes <= self._current_training_processes:
-            raise MaxTrainingError
+        if self._pool_processes <= self._current_pool_processes:
+            raise MaxPoolProcessError
 
         if project in self.project_store:
             self.project_store[project].status = STATUS_TRAINING
@@ -322,10 +324,10 @@ class DataRouter(object):
         def training_callback(model_path):
             model_dir = os.path.basename(os.path.normpath(model_path))
             self.project_store[project].update(model_dir)
-            self._current_training_processes -= 1
-            self.project_store[project].current_training_processes -= 1
+            self._current_pool_processes -= 1
+            self.project_store[project].current_pool_processes -= 1
             if (self.project_store[project].status == STATUS_TRAINING and
-                    self.project_store[project].current_training_processes ==
+                    self.project_store[project].current_pool_processes ==
                     0):
                 self.project_store[project].status = STATUS_READY
             return model_path
@@ -333,8 +335,8 @@ class DataRouter(object):
         def training_errback(failure):
             logger.warning(failure)
 
-            self._current_training_processes -= 1
-            self.project_store[project].current_training_processes -= 1
+            self._current_pool_processes -= 1
+            self.project_store[project].current_pool_processes -= 1
             self.project_store[project].status = STATUS_FAILED
             self.project_store[project].error_message = str(failure)
 
@@ -342,8 +344,8 @@ class DataRouter(object):
 
         logger.debug("New training queued")
 
-        self._current_training_processes += 1
-        self.project_store[project].current_training_processes += 1
+        self._current_pool_processes += 1
+        self.project_store[project].current_pool_processes += 1
 
         result = self.pool.submit(do_train_in_worker,
                                   train_config,
@@ -365,10 +367,11 @@ class DataRouter(object):
                  model: Optional[Text] = None) -> Deferred:
         """Perform a model evaluation."""
 
-        logger.debug("Evaluation request received for project '{}' and model '{}'.".format(project, model))
+        logger.debug("Evaluation request received for "
+                     "project '{}' and model '{}'.".format(project, model))
 
-        if self._training_processes <= self._current_training_processes:
-            raise MaxTrainingError
+        if self._pool_processes <= self._current_pool_processes:
+            raise MaxPoolProcessError
 
         project = project or RasaNLUModelConfig.DEFAULT_PROJECT_NAME
         data_path = utils.create_temporary_file(data, "_training_data")
@@ -380,38 +383,36 @@ class DataRouter(object):
         model = model or self.project_store[project]._dynamic_load_model(model)
 
         if model == FALLBACK_MODEL_NAME:
-            raise InvalidProjectError("No model in project '{}' to "
-                                      "evaluate.".format(project))
+            raise UnsupportedModelError("No model in project '{}' to "
+                                        "evaluate.".format(project))
 
         model_path = os.path.join(self.project_store[project]._path, model)
 
         def evaluation_callback(result):
             logger.debug("Evaluation was successful")
 
-            self._current_training_processes -= 1
-            self.project_store[project].current_training_processes -= 1
+            self._current_pool_processes -= 1
+            self.project_store[project].current_pool_processes -= 1
 
             return result
 
         def evaluation_errback(failure):
             logger.warning(failure)
 
-            self._current_training_processes -= 1
-            self.project_store[project].current_training_processes -= 1
+            self._current_pool_processes -= 1
+            self.project_store[project].current_pool_processes -= 1
 
             return failure
 
         logger.debug("New evaluation queued.")
 
-        self._current_training_processes += 1
-        self.project_store[project].current_training_processes += 1
+        self._current_pool_processes += 1
+        self.project_store[project].current_pool_processes += 1
 
         result = self.pool.submit(run_evaluation,
                                   data_path,
                                   model_path,
-                                  None,
-                                  None,
-                                  None)
+                                  errors_filename=None)
 
         result = deferred_from_future(result)
         result.addCallback(evaluation_callback)