+++ rasa/nlu/project.py
@@ -161,7 +161,7 @@ class Project(object):
         self._component_builder = component_builder
         self._models = {}
         self.status = STATUS_READY
-        self.current_training_processes = 0
+        self.current_pool_processes = 0
         self._reader_lock = Lock()
         self._loader_lock = Lock()
         self._writer_lock = Lock()
@@ -391,7 +391,7 @@ class Project(object):
         return {
             "status": status,
             "error_message": error_message,
-            "current_training_processes": self.current_training_processes,
+            "current_pool_processes": self.current_pool_processes,
             "available_models": list(self._models.keys()),
             "loaded_models": self._list_loaded_models()
         }