+++ rasa/nlu/server.py
@@ -14,7 +14,7 @@ from rasa.nlu import config, utils
 import rasa.nlu.cli.server as cli
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.data_router import (
-    DataRouter, InvalidProjectError, MaxTrainingError)
+    DataRouter, InvalidProjectError, MaxPoolProcessError, UnsupportedModelError)
 from rasa.constants import MINIMUM_COMPATIBLE_VERSION
 from rasa.nlu.train import TrainingException
 from rasa.nlu.utils import json_to_string, read_endpoints
@@ -287,7 +287,7 @@ class RasaNLU(object):
             zip_content = io.open(zipped_path, 'r+b').read()
             return returnValue(zip_content)
 
-        except MaxTrainingError as e:
+        except MaxPoolProcessError as e:
             request.setResponseCode(403)
             returnValue(json_to_string({"error": "{}".format(e)}))
         except InvalidProjectError as e:
@@ -316,10 +316,13 @@ class RasaNLU(object):
                                                        params.get('project'),
                                                        params.get('model'))
             returnValue(json_to_string(response))
-        except MaxTrainingError:
+        except MaxPoolProcessError as e:
             request.setResponseCode(403)
-            returnValue("An idle training process is required to evaluate")
+            returnValue(json_to_string({"error": "{}".format(e)}))
         except InvalidProjectError as e:
+            request.setResponseCode(500)
+            returnValue(json_to_string({"error": "{}".format(e)}))
+        except UnsupportedModelError as e:
             request.setResponseCode(500)
             returnValue(json_to_string({"error": "{}".format(e)}))
         except Exception as e: